Move oidpath and oid.logopath variables to main fK.cf
This move was mainly inspired by the fact that the google closure
compiler simply erased oid.oidpath since it did not "see" the variable
anywhere else. We may still not be able to use even the
SIMPLE_OPTIMIZATIONS but it seemed like the problem was a hint about
better practices.