Separate facebook snippet method into 2 methods: script elem and button
Eventually we will probably drop this altogether, as the FB stuff is
too vital to be so removed from what is going on. At first I was
treating this as a black-box, cargo code element but I think that both
the button and the script will need to be more carefully integrated
into the page and the javascript.