Support for tagPage in fkTag and tagdex.
This is, at least symbolically, a major shift away from a separate web
interface and towards using the basic (formerly) tag.php and
resource.php functions to generate as much content as possible through
XSLT stylesheets.

One advantage so far is that fabula specific boilerplate can be
included via specific stylesheets, rather than cluttering up code in
general.

There will probably be limits to this approach in some cases where it
will just be simpler to code up a page, but in general it seems like a
clean solution.