folksoRights: functions for deleting
We are working on the admin.php function for updating user rights. The
interface simplifies things and hides some of the complexity of the
rights system. The notion of "service" disappears for the user. This
could be changed later if need be but for now there is no reason to
make things more complicated. So for the rights we needed to be able
to say "make sure this user doesn't have any rights above level X".

rightsAsArray()
removeRightsAbove()
maxRight()