* js/pageinit.js: Toggle 'loggedIn' trigger
Just an attempt, not entirely sure this will do exactly what is
intended, but the idea is that this won't fire unless we are going
from not logged in to logged in. One would have to check how
fK.loginStatus gets modified at different times to be sure that this
really works.