* js/folkso-user.js, js/pageinit.js: tinymce init moved to function
There is a bug on login, because the tinymce is initialized in a box
that isn't visible yet if the user hasn't logged in yet. The idea is
to incorporate the tinymce init process into the loggedIn event,
rather than page load. This isn't quite perfect yet, because there is
a bug in the event system that needs to be fixed first, before
everything works together.