IE8 bug: hack to avoid alert messages
For some reason, in IE8, jquery-ui tabs, when opened, cause
$(document).ready to fire. This doesn't seem to matter much, except
when logging out, because all of the logging in events are fired
again. The result, in IE8, was that the alerts for failed ajax
interactions were all firing on logout. I couldn't find why any of
this was happening so, ungloriously, I just suppressed the alerts for
IE.